GuidedBugEntry more bug fixes and tweaks

+++ enter_bug.cgi
@@ -68,12 +68,14 @@ Bugzilla::Hook::process('enter_bug_start', { vars => $vars });
 # All pages point to the same part of the documentation.
 $vars->{'doc_section'} = 'bugreports.html';
 
-# Purpose: force guided format for newbies
-$cgi->param(-name=>'format', -value=>'guided') 
-    if !$cgi->param('format') && !$user->in_group('canconfirm');
+if (!$vars->{'disable_guided'}) {
+    # Purpose: force guided format for newbies
+    $cgi->param(-name=>'format', -value=>'guided') 
+        if !$cgi->param('format') && !$user->in_group('canconfirm');
 
-$cgi->delete('format') 
-    if ($cgi->param('format') && ($cgi->param('format') eq "__default__"));
+    $cgi->delete('format') 
+        if ($cgi->param('format') && ($cgi->param('format') eq "__default__"));
+}
 
 my $product_name = trim($cgi->param('product') || '');
 # Will contain the product object the bug is created in.
